+++ src/core/BaseController.php
@@ -17,7 +17,6 @@ declare(strict_types=1);
 
 namespace herosphp\core;
 
-use app\controller\UserValidate;
 use herosphp\annotation\Validate;
 use herosphp\utils\StringUtil;
 use herosphp\vo\JsonVo;
@@ -48,7 +47,6 @@ abstract class BaseController extends Template
         foreach ($reflectionAttributes as $validAttribute) {
             /** @var Validate $methodValidInstance */
             $methodValidInstance = $validAttribute->newInstance();
-            /** @var UserValidate $methodVInstance*/
             $methodVInstance = new ($methodValidInstance->class);
             if (! $methodVInstance instanceof \herosphp\validate\Validate) {
                 continue;